refactor: remove macros.rs (import_all) and re-organise
