+++ rsrl/src/control/ac/mod.rs
@@ -1,6 +1,16 @@
-import_all!(cacla);
-import_all!(qac);
-import_all!(tdac);
-import_all!(a2c);
-import_all!(nac);
-import_all!(offpac);
+//! Actor-critic algorithms.
+pub mod cacla;
+pub mod qac;
+pub mod tdac;
+pub mod a2c;
+pub mod nac;
+pub mod offpac;
+
+pub use self::{
+    cacla::CACLA,
+    qac::QAC,
+    tdac::TDAC,
+    a2c::A2C,
+    nac::NAC,
+    offpac::OffPAC,
+};
