--- rsrl/src/control/gtd/mod.rs
@@ -1,1 +1,0 @@
-import_all!(greedy_gq);
