+++ rsrl/src/control/mc/mod.rs
@@ -1,2 +1,8 @@
-import_all!(reinforce);
-import_all!(baseline_reinforce);
+//! Monte-Carlo policy gradient algorithms.
+pub mod reinforce;
+pub mod baseline_reinforce;
+
+pub use self::{
+    reinforce::REINFORCE,
+    baseline_reinforce::BaselineREINFORCE,
+};
