+++ rsrl/src/control/td/mod.rs
@@ -1,13 +1,27 @@
+//! Temporal-difference control algorithms.
 // Off-policy:
-import_all!(q_learning);
-import_all!(q_lambda);
-import_all!(q_sigma);
-import_all!(pal);
+pub mod q_learning;
+pub mod q_lambda;
+pub mod q_sigma;
+pub mod pal;
+
+pub use self::{
+    q_learning::QLearning,
+    q_lambda::QLambda,
+    q_sigma::QSigma,
+    pal::PAL,
+};
 
 // On-policy:
-import_all!(sarsa);
-import_all!(sarsa_lambda);
-import_all!(expected_sarsa);
+pub mod sarsa;
+pub mod sarsa_lambda;
+pub mod expected_sarsa;
+
+pub use self::{
+    sarsa::SARSA,
+    sarsa_lambda::SARSALambda,
+    expected_sarsa::ExpectedSARSA,
+};
 
 // TODO:
 // PQ(lambda) - http://proceedings.mlr.press/v32/sutton14.pdf
