--- rsrl/src/control/totd/mod.rs
@@ -1,5 +1,0 @@
-// Off-policy:
-import_all!(to_q_lambda);
-
-// On-policy:
-import_all!(to_sarsa_lambda);
