--- rsrl/src/prediction/gtd/mod.rs
@@ -1,11 +1,0 @@
-import_all!(tdc);
-import_all!(gtd2);
-// import_all!(gtd_lambda_is);
-
-// TODO:
-// ABQ(lambda) - https://arxiv.org/pdf/1702.03006.pdf
-// GQ(lambda) - http://agi-conf.org/2010/wp-content/uploads/2009/06/paper_21.pdf
-// GTD(lambda) - https://era.library.ualberta.ca/files/8s45q967t/Hamid_Maei_PhDThesis.pdf
-// True online GTD(lambda) - http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.487.2451&rep=rep1&type=pdf
-// GTD2(lambda)-MP - https://arxiv.org/pdf/1602.08771.pdf
-// TDC(lambda)-MP - https://arxiv.org/pdf/1602.08771.pdf
