+++ rsrl/src/prediction/td/gtd2.rs
